requirements.txt
# API
flask
requests

# TEXT TO SPEECH
tts
torch
torchaudio

# LANGUAGE MODEL
ollama
newspaper3k
ddgs
# MACHINE LEARNING

# OTHER
coloredlogs